requirements.txt
google-generativeai==0.6.0
openai==1.35.15
python-dotenv==1.0.0
chromadb==0.4.24
pydantic==2.8.2
requests==2.32.3
mss==9.0.1
Pillow==10.4.0
chainlit